// src/types/mapbox.ts
export interface MapboxSuggestion {
  name: string;
  name_preferred?: string;
  mapbox_id: string;
  feature_type: string;
  address?: string;
  full_address?: string;
  place_formatted: string;
  context: {
    country?: {
      id?: string;
      name: string;
      country_code: string;
      country_code_alpha_3: string;
    };
    region?: {
      id?: string;
      name: string;
      region_code: string;
      region_code_full: string;
    };
    postcode?: {
      id?: string;
      name: string;
    };
    district?: {
      id?: string;
      name: string;
    };
    place?: {
      id?: string;
      name: string;
    };
    locality?: {
      id?: string;
      name: string;
    };
    neighborhood?: {
      id?: string;
      name: string;
    };
    address?: {
      id?: string;
      name: string;
      address_number?: string;
      street_name?: string;
    };
    street?: {
      id?: string;
      name: string;
    };
  };
  language: string;
  maki?: string;
  poi_category?: string[];
  poi_category_ids?: string[];
  brand?: string[];
  brand_id?: string[];
  external_ids?: Record<string, string>;
  metadata?: Record<string, unknown>;
  distance?: number;
  eta?: number;
  added_distance?: number;
  added_time?: number;
}